ref: Move RestConfig to the ktor package.
