+++ jicofo/src/main/kotlin/org/jitsi/jicofo/JicofoServices.kt
@@ -37,7 +37,7 @@ import org.jitsi.jicofo.jibri.JibriDetectorMetrics
 import org.jitsi.jicofo.metrics.GlobalMetrics
 import org.jitsi.jicofo.metrics.JicofoMetricsContainer
 import org.jitsi.jicofo.rest.Application
-import org.jitsi.jicofo.rest.RestConfig
+import org.jitsi.jicofo.ktor.RestConfig
 import org.jitsi.jicofo.util.SynchronizedDelegate
 import org.jitsi.jicofo.version.CurrentVersionImpl
 import org.jitsi.jicofo.xmpp.XmppServices
