+++ jicofo/src/main/kotlin/org/jitsi/jicofo/ktor/RestConfig.kt
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jitsi.jicofo.rest
+package org.jitsi.jicofo.ktor
 
 import org.jitsi.config.JitsiConfig
 import org.jitsi.metaconfig.config
